requirements.txt
streamlit
python-dotenv
supabase
qrcode
Pillow
